+++ src/main/java/nz/ac/auckland/se306/group12/models/Node.java
@@ -19,7 +19,7 @@ public class Node {
   @EqualsAndHashCode.Include
   private final String label;
   @EqualsAndHashCode.Include
-  private final long weight;
+  private final int weight;
   private final Set<Edge> incomingEdges = new HashSet<>();
   private final Set<Edge> outgoingEdges = new HashSet<>();
